% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{checkPara}
\alias{checkPara}
\title{Check parameters for functions}
\usage{
checkPara(paralst, auto = T)
}
\arguments{
\item{paralst}{parameters list}
}
\value{
Validate parameters in a ENV
}
\description{
Check parameters for functions
}
\author{
Qiong Zhang
}
