% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{fetchFeasMeta}
\alias{fetchFeasMeta}
\title{Fetch features and metadata}
\usage{
fetchFeasMeta(
  seu,
  features,
  slot = "data",
  assay = "RNA",
  keywords = NULL,
  cell.ident = T
)
}
\arguments{
\item{seu}{Seurat object}

\item{features}{which feature(s)}

\item{slot}{target slot in assay}

\item{assay}{target assay in seu object}

\item{keywords}{A list corresponds to the feature parameter.}
}
\value{
A data table fits the value
}
\description{
This function fetches features and metadata based on the key words.
}
\examples{
\dontrun{
# Load example data and gene sets
filter <- list(seurat_annotations = c('Memory CD4 T', 'Naive CD4 T'), CD4 = c('>1', '<1.6'))

fetchFeasMeta(seu = pbmc3k.final, features = c("seurat_annotations", 'CD4'), keywords = filter)
}

}
\author{
Qiong Zhang
}
