% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{fetchSetsMsigdb}
\alias{fetchSetsMsigdb}
\title{Fetch msigdb gene sets}
\usage{
fetchSetsMsigdb(
  species = "Homo sapiens",
  category = "H",
  subcategory = NULL,
  formatId = "symbol",
  localdb = F
)
}
\arguments{
\item{seu}{Seurat object}

\item{features}{which feature(s)}

\item{keywords}{A list corresponds to the feature parameter.}

\item{assay}{target assay in seu object}

\item{slot}{target slot in assay}
}
\value{
A data table fits the value
}
\description{
This function fetches getsets from msigdb.
}
\examples{
\dontrun{
# Load example data and gene sets
filter <- list(seurat_annotations = c('Memory CD4 T', 'Naive CD4 T'), CD4 = c('>1', '<1.6'))

fetchFeasMeta(seu = pbmc3k.final, features = c("seurat_annotations", 'CD4'), keywords = filter)
}

}
\author{
Qiong Zhang
}
