% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{geneset2w}
\alias{geneset2w}
\title{Construct a weight vector or data.frame for gene sets}
\usage{
geneset2w(genesets, outformat = "l")
}
\arguments{
\item{genesets}{Seurat object or expression matrix}

\item{outformat}{'d' (data.frame) or l (list)}
}
\value{
weight for genesets
}
\description{
Construct a weight vector or data.frame for gene sets
}
\author{
Qiong Zhang
}
