% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.common.R
\name{get_random_loss_sparse_matrix}
\alias{get_random_loss_sparse_matrix}
\title{Random drop n percent of expressed gene for cells and return the matrix}
\usage{
get_random_loss_sparse_matrix(m, n)
}
\arguments{
\item{m}{Expression matrix of gene(row) X cell (column)}

\item{n}{The percentage of genes should be dropped}
}
\value{
Expression matrix with dropped genes
}
\description{
Random drop n percent of expressed gene for cells and return the matrix
}
\author{
Qiong Zhang
}
