% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{getPosNegSets}
\alias{getPosNegSets}
\alias{splitGeneSet2PosNeg}
\title{Extract positive and negative genes and seperate them to 2 lists from genesets}
\usage{
splitGeneSet2PosNeg(genesets, geneall, ignore = T, discard = F)
}
\arguments{
\item{genesets}{A list of gene sets to be used for gene signature scoring.}
}
\value{
gene list contains 2 elements: positive and negative genesets.
}
\description{
This function seperates genes to 2 parts, positive and negative parts, based on their suffix label
}
\examples{
\dontrun{
# Load example data and gene sets
genesets <- list(CD4 = c("CD3D", "CD3E", "CD4", "CD8A-", "CD8B-"), B = c("CD19+", "CD79A", "CD3D-"))

genesets.pn <- getPosNegSets(genesets = genesets)
}

}
\author{
Qiong Zhang
}
