% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GScores}
\alias{GScores}
\title{Calculate Signature score for genes from single-cell data}
\usage{
GScores(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = c("GSignatures"),
  scale = F,
  return.type = "obj"
)
}
\arguments{
\item{object}{Seurat object.}

\item{genesets}{A list of genesets, for example:
\code{list(CD8CTL = c("CD3D", "CD3E", "CD4-", "CD8A", "CD8B", "KLRF1", "GZMA", "CCL5"),
Bcell = c("CD19", "CD79A", "CD3D-", "CD3E-"))}
Users can also specify genes with negative effects by suffixing the '-', see examples of CD8CTL}

\item{assay}{The sce object assay where the data is to be found}

\item{slot}{Which slot used in the analysis}

\item{methods}{methods used to calculate gene set signatures}

\item{scale}{scale gene set signatures (from 0 to 1)}
}
\description{
Calculate Signature score for genes from single-cell data
}
\examples{
library(GScores)
genesets <- list(CD8CTL = c("CD3D", "CD3E", "CD4-", "CD8A", "CD8B", "KLRF1", "GZMA", "CCL5"),
    Bcell = c("CD19", "CD79A", "CD3D-", "CD3E-"))
methods <-  c('Gsignatures', 'AUCell', 'UCell', 'AddModuleScore','ssgsea', 'VAM', 'plage', 'pagoda2', 'VISION')
scores <- GScores(pbmc3k.final, genesets, assay = 'RNA', slot = 'counts', methods = methods, scale = T)
head(scores)
}
