% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.common.R
\name{list2SparseMat}
\alias{list2SparseMat}
\title{Transfer the sparseVector list to Matrix format}
\usage{
list2SparseMat(list)
}
\arguments{
\item{list}{sparseVector list}
}
\value{
Sparse matrix
}
\description{
Transfer the sparseVector list to Matrix format
}
\author{
Qiong Zhang
}
