% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_AddModuleScore}
\alias{run_AddModuleScore}
\title{Run Seurat scoring on a Seurat object}
\usage{
run_AddModuleScore(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "AddModuleScore",
  seed = 10086,
  ncores = 1
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for Seurat scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the Seurat scores added as a new assay called "Seurat".
}
\description{
This function performs Seurat scoring on a given Seurat object using the specified gene sets.
It calculates the Seurat scores and stores them in a new assay called "Seurat".
}
\author{
Qiong Zhang
}
