% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_AUCell}
\alias{run_AUCell}
\title{Run AUCell analysis on a Seurat object or expression matrix}
\usage{
run_AUCell(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "AUCell",
  aucratio = 0.05
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data or an expression matrix.}

\item{genesets}{A list of gene sets to be used for AUCell scoring.}

\item{aucratio}{A numeric value specifying the ratio for determining the maximum rank in AUCell scoring.}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object or expression matrix with the AUCell scores added as a new assay called "AUCell".
}
\description{
This function performs AUCell analysis on a given Seurat object or expression matrix using the specified gene sets.
It calculates the AUCell scores and stores them in a new assay called "AUCell".
}
\examples{
\dontrun{
# Load example data and gene sets
library(AUCell)
pbmc_small <- Seurat::pbmc_small()
example_genesets <- list(set1 = c("CD3D", "CD3E"), set2 = c("CD19", "CD79A"))

# Run AUCell analysis
pbmc_small_aucell <- run_AUCell(pbmc_small, genesets = example_genesets, aucratio = 0.05)
}

}
\author{
Qiong Zhang
}
