% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_decoupleR}
\alias{run_decoupleR}
\title{Run several methods in decoupleR package to perform gene set scoring on a Seurat object}
\usage{
run_decoupleR(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "viper",
  scale = T
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for several scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the several scores added as a new assay named by related method.
}
\description{
This function performs several scoring on a given Seurat object using the specified gene sets.
It calculates the several scores and stores them in a new assay called "several".
}
\author{
Qiong Zhang
}
