% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_GSignatures}
\alias{run_GSignatures}
\title{Run GSignatures scoring on a Seurat object}
\usage{
run_GSignatures(object, genesets, assay = "RNA", slot = "counts")
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for GSignatures scoring}
}
\value{
The input Seurat object with the GSignatures scores added as a new assay called "GSignatures".
}
\description{
This function performs GSignatures scoring on a given Seurat object using the specified gene sets.
It calculates the GSignatures scores and stores them in a new assay called "GSignatures".
}
\author{
Qiong Zhang
}
