% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_GSXX}
\alias{run_GSXX}
\title{Run ssGSEA or c("gsva", "zscore", "plage") analysis on a Seurat object}
\usage{
run_GSXX(
  object,
  genesets,
  assay = "RNA",
  slot = "data",
  methods = "ssgsea",
  kcdf = "Gaussian",
  ssgsea.norm = F
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for ssGSEA scoring.}

\item{methods}{could be one of c("gsva", "ssgsea", "zscore", "plage")}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the ssGSEA scores added as a new assay called "ssgsea".
}
\description{
This function performs ssGSEA (single sample Gene Set Enrichment Analysis) on a given Seurat object using the specified gene sets.
It calculates the ssGSEA scores and stores them in a new assay called "ssgsea".
}
\examples{
\dontrun{
# Load example data and gene sets
library(Seurat)
pbmc_small <- Seurat::pbmc_small()
example_genesets <- list(set1 = c("CD3D", "CD3E"), set2 = c("CD19", "CD79A"))

# Run ssGSEA analysis
pbmc_small_ssgsea <- runssGSEA(pbmc_small, genesets = example_genesets, aucratio = 0.05)
}

}
\author{
Qiong Zhang
}
