% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_pagoda2}
\alias{run_pagoda2}
\title{Run pagoda2 scoring on a Seurat object}
\usage{
run_pagoda2(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "pagoda2",
  seed = 10086,
  ncores = 1
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for pagoda2 scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the pagoda2 scores added as a new assay called "pagoda2".
}
\description{
This function performs pagoda2 scoring on a given Seurat object using the specified gene sets.
It calculates the pagoda2 scores and stores them in a new assay called "pagoda2".
}
\author{
Qiong Zhang
}
