% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_scGSEA}
\alias{run_scGSEA}
\title{Run scGSEA scoring on a Seurat object}
\usage{
run_scGSEA(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "scGSEA",
  ncores = 1
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for scGSEA scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the scGSEA scores added as a new assay called "scGSEA".
}
\description{
This function performs scGSEA scoring on a given Seurat object using the specified gene sets.
It calculates the scGSEA scores and stores them in a new assay called "scGSEA".
}
\author{
Qiong Zhang
}
