% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_scSE}
\alias{run_scSE}
\title{Run scSE scoring on a Seurat object}
\usage{
run_scSE(object, genesets, assay = "RNA", slot = "counts", methods = "scSE")
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for scSE scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the scSE scores added as a new assay called "scSE".
}
\description{
This function performs scSE scoring on a given Seurat object using the specified gene sets.
It calculates the scSE scores and stores them in a new assay called "scSE".
}
\author{
Qiong Zhang
}
