% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_singscore}
\alias{run_singscore}
\title{Run singscore analysis on a Seurat object}
\usage{
run_singscore(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "singscore"
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for singscore analysis.}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the singscore scores added as a new assay called "singscore".
}
\description{
This function performs singscore analysis on a given Seurat object using the specified gene sets.
It calculates the singscore scores and stores them in a new assay called "singscore".
}
\examples{
\dontrun{
# Load example data and gene sets
library(Seurat)
pbmc_small <- Seurat::pbmc_small()
example_genesets <- list(set1 = c("CD3D", "CD3E"), set2 = c("CD19", "CD79A"))

# Run singscore analysis
pbmc_small_singscore <- runsingscore(pbmc_small, genesets = example_genesets, aucratio = 0.05)
}

}
\author{
Qiong Zhang
}
