% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_UCell}
\alias{run_UCell}
\title{Run UCell analysis on a Seurat object}
\usage{
run_UCell(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "UCell",
  MaxRank = 1500
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for UCell scoring.}

\item{MaxRank}{Integer specifying the maximum rank for UCell scoring. Default is 1500.}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the UCell scores added as a new assay called "UCell".
}
\description{
This function performs UCell analysis on a given Seurat object using the specified gene sets.
It calculates the UCell scores and stores them in a new assay called "UCell".
}
\examples{
\dontrun{
# Load example data and gene sets
library(Seurat)
pbmc_small <- Seurat::pbmc_small()
example_genesets <- list(set1 = c("CD3D", "CD3E"), set2 = c("CD19", "CD79A"))

# Run UCell analysis
pbmc_small_ucell <- runUCell(pbmc_small, genesets = example_genesets)
}


}
\author{
Qiong Zhang
}
