% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_VAM}
\alias{run_VAM}
\title{Run VAM scoring on a Seurat object}
\usage{
run_VAM(object, genesets, assay = "RNA", slot = "data", methods = "VAM")
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for VAM scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the VAM scores added as a new assay called "VAM".
}
\description{
This function performs VAM scoring on a given Seurat object using the specified gene sets.
It calculates the VAM scores and stores them in a new assay called "VAM".
}
\author{
Qiong Zhang
}
