% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScores.R
\name{run_VISION}
\alias{run_VISION}
\title{Run VISION scoring on a Seurat object}
\usage{
run_VISION(
  object,
  genesets,
  assay = "RNA",
  slot = "counts",
  methods = "VISION"
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data.}

\item{genesets}{A list of gene sets to be used for VISION scoring}

\item{ncores}{Integer specifying the number of cores to be used for parallel computation. Default is 1.}
}
\value{
The input Seurat object with the VISION scores added as a new assay called "VISION".
}
\description{
This function performs VISION scoring on a given Seurat object using the specified gene sets.
It calculates the VISION scores and stores them in a new assay called "VISION".
}
\author{
Qiong Zhang
}
