% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/util.functions.R
\name{scaleXY}
\alias{scaleXY}
\title{rescale a vector between 0 (lowest) and 1 (highest)}
\usage{
scaleXY(x, y)

scaleXY(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector with 2 values with up and down boundaries}
}
\value{
rescale vector with value between 0 and 1

rescale vector with value between y[2] and y[1]
}
\description{
rescale a vector between 0 (lowest) and 1 (highest)

rescale a vector between 0 (lowest) and 1 (highest)
}
\author{
Qiong Zhang
}
