% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.common.R
\name{simSparseMat}
\alias{simSparseMat}
\title{Simulate the matrix based on the sparsity}
\usage{
simSparseMat(Mat, prop = 0.1, seed = 10086, addname = T, merge = T, random = T)
}
\arguments{
\item{Mat}{Expression or binary matrix}

\item{prop}{The ratio of simulation cells}

\item{seed}{Seed used}

\item{addname}{Whether to add the prefix of 'pseudo' to simulation cells}

\item{merge}{Whether to merge with original matrix}

\item{random}{Use global sparsity or random select the cell sparsity}
}
\value{
Simulation matrix with similar sparsity as original one
}
\description{
Simulate the matrix based on the sparsity
}
\author{
Qiong Zhang
}
