% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.functions.R
\name{vldMethods}
\alias{vldMethods}
\title{Validate whether users had installed selected methods}
\usage{
vldMethods(method.use, install = T)
}
\arguments{
\item{expM}{Seurat object or expression matrix}

\item{geneSets}{which feature(s)}
}
\value{
GeneSets contain genes in the expression matrix
}
\description{
Validate whether users had installed selected methods
}
\author{
Qiong Zhang
}
